Fix issue with phantom devices & ZoneGroupMember not being an array
If a device was unplugged or previously part of the Sonos Network, it
comes through getAllGroups, but has no device associated. This would
cause _parseZoneGroups to fail.

Also, groups.ZoneGroupMember can sometimes be an Object and not an array
apparently, although I can't figure out how this is possible based
on the node-sonos library. It is coerced to an array there. But, one user
did experience this, although I can't reproduce it. Now, we coerce it
to an array also. Hope this fixes the issue.