Fix socketio issue that impacted production build
Clients that were not on the same host server as the Sonos Web app,
were unable to connect because we specified localhost as the socketio
server. The clients were then trying to communicate with their own localhost
instead of communicating with the actual server. Changing the variable
to an empty string, allows socketio to connect the client to the proper
place.